Conflicts:
	src/main/java/com/afd/service/user/IUserService.java

+++ src/main/java/com/afd/model/product/BcSpecValue.java
@@ -5,21 +5,21 @@ import java.io.Serializable;
 public class BcSpecValue implements Serializable {
 	private static final long serialVersionUID = 8376251680755359452L;
 
-	private Integer bcSvId;
+	private Long bcSvId;
 
-    private Integer bcSpecId;
+    private Long bcSpecId;
 
-    private Integer specValueId;
+    private Long specValueId;
 
     private String specValueName;
 
     private String imgUrl;
 
-    private Byte displayOrder;
+    private Integer displayOrder;
 
     private String status;
 
-    private Short bcId;
+    private Integer bcId;
 
     private Boolean isFilter;
 
@@ -27,91 +27,92 @@ public class BcSpecValue implements Serializable {
 
     private Integer mDisplayPosition;
 
-    public Integer getBcSvId() {
-        return bcSvId;
-    }
+	public Long getBcSvId() {
+		return bcSvId;
+	}
 
-    public void setBcSvId(Integer bcSvId) {
-        this.bcSvId = bcSvId;
-    }
+	public void setBcSvId(Long bcSvId) {
+		this.bcSvId = bcSvId;
+	}
 
-    public Integer getBcSpecId() {
-        return bcSpecId;
-    }
+	public Long getBcSpecId() {
+		return bcSpecId;
+	}
 
-    public void setBcSpecId(Integer bcSpecId) {
-        this.bcSpecId = bcSpecId;
-    }
+	public void setBcSpecId(Long bcSpecId) {
+		this.bcSpecId = bcSpecId;
+	}
 
-    public Integer getSpecValueId() {
-        return specValueId;
-    }
+	public Long getSpecValueId() {
+		return specValueId;
+	}
 
-    public void setSpecValueId(Integer specValueId) {
-        this.specValueId = specValueId;
-    }
+	public void setSpecValueId(Long specValueId) {
+		this.specValueId = specValueId;
+	}
 
-    public String getSpecValueName() {
-        return specValueName;
-    }
+	public String getSpecValueName() {
+		return specValueName;
+	}
 
-    public void setSpecValueName(String specValueName) {
-        this.specValueName = specValueName;
-    }
+	public void setSpecValueName(String specValueName) {
+		this.specValueName = specValueName;
+	}
 
-    public String getImgUrl() {
-        return imgUrl;
-    }
+	public String getImgUrl() {
+		return imgUrl;
+	}
 
-    public void setImgUrl(String imgUrl) {
-        this.imgUrl = imgUrl;
-    }
+	public void setImgUrl(String imgUrl) {
+		this.imgUrl = imgUrl;
+	}
 
-    public Byte getDisplayOrder() {
-        return displayOrder;
-    }
+	public Integer getDisplayOrder() {
+		return displayOrder;
+	}
 
-    public void setDisplayOrder(Byte displayOrder) {
-        this.displayOrder = displayOrder;
-    }
+	public void setDisplayOrder(Integer displayOrder) {
+		this.displayOrder = displayOrder;
+	}
 
-    public String getStatus() {
-        return status;
-    }
+	public String getStatus() {
+		return status;
+	}
 
-    public void setStatus(String status) {
-        this.status = status;
-    }
+	public void setStatus(String status) {
+		this.status = status;
+	}
 
-    public Short getBcId() {
-        return bcId;
-    }
+	public Integer getBcId() {
+		return bcId;
+	}
 
-    public void setBcId(Short bcId) {
-        this.bcId = bcId;
-    }
+	public void setBcId(Integer bcId) {
+		this.bcId = bcId;
+	}
 
-    public Boolean getIsFilter() {
-        return isFilter;
-    }
+	public Boolean getIsFilter() {
+		return isFilter;
+	}
 
-    public void setIsFilter(Boolean isFilter) {
-        this.isFilter = isFilter;
-    }
+	public void setIsFilter(Boolean isFilter) {
+		this.isFilter = isFilter;
+	}
 
-    public Boolean getIsMobileDisplay() {
-        return isMobileDisplay;
-    }
+	public Boolean getIsMobileDisplay() {
+		return isMobileDisplay;
+	}
 
-    public void setIsMobileDisplay(Boolean isMobileDisplay) {
-        this.isMobileDisplay = isMobileDisplay;
-    }
+	public void setIsMobileDisplay(Boolean isMobileDisplay) {
+		this.isMobileDisplay = isMobileDisplay;
+	}
 
-    public Integer getmDisplayPosition() {
-        return mDisplayPosition;
-    }
+	public Integer getmDisplayPosition() {
+		return mDisplayPosition;
+	}
 
-    public void setmDisplayPosition(Integer mDisplayPosition) {
-        this.mDisplayPosition = mDisplayPosition;
-    }
+	public void setmDisplayPosition(Integer mDisplayPosition) {
+		this.mDisplayPosition = mDisplayPosition;
+	}
+    
 }
